+++ src/components/Forms/PageBlockForm.module.scss
@@ -1,0 +1,77 @@
+.container {
+    width: 100%;
+    height: auto;
+    display: flex;
+  }
+  
+  .addImgButton {
+    width: 100%;
+    display: flex;
+    cursor: pointer;
+    font-size: 2rem;
+    font-weight: 100;
+    height: 5.3125rem;
+    align-items: center;
+    border-radius: .375rem;
+    justify-content: center;
+    border: solid .0625rem #000000;
+  }
+  
+  .formBox {
+    display: flex;
+    width: 100%;
+    flex-direction: column;
+    justify-content: space-between;
+  
+    .forms {
+      padding: 0 20px 0 20px;
+    }
+  
+    .formText {
+        margin-top: 30px;
+    }
+  }
+  
+  input[type="text"] {
+    display: flex;
+    width: 100%;
+    height: 3.375rem;
+    font-size: 1rem;
+    border-radius: .375rem;
+    padding-left: .875rem;
+    border: solid .0625rem #e2e6ef;
+    &::placeholder {
+      color: #e2e6ef;
+    }
+  }
+  
+  .thumbnailImg {
+    display: flex;
+    cursor: pointer;
+    overflow: hidden;
+    width: 5.3125rem;
+    height: 5.3125rem;
+    position: relative;
+    border-radius: .25rem;
+    img {
+      width: auto;
+      height: auto;
+    }
+    .minus {
+      top: 0;
+      left: 0;
+      right: 0;
+      bottom: 0;
+      display: none;
+      position: absolute;
+      align-items: center;
+      justify-content: center;
+    }
+  }
+  
+  .thumbnailImg:hover {
+    .minus {
+      display: flex;
+      color: #ff0606;
+    }
+  }
